+++ src/lib/ui/Home.svelte
@@ -1,11 +1,9 @@
 <script lang="ts">
 	import type {Readable} from 'svelte/store';
-	import {throttle} from 'throttle-debounce';
 
 	import type {Space} from '$lib/vocab/space/space.js';
 	import Avatar from '$lib/ui/Avatar.svelte';
-	import EntityIcon from '$lib/ui/EntityIcon.svelte';
-	import HueInput from '$lib/ui/HueInput.svelte';
+	import CommunityHue from '$lib/ui/CommunityHue.svelte';
 	import SpaceInfo from '$lib/ui/SpaceInfo.svelte';
 	import {getApp} from '$lib/ui/app';
 	import type {Community} from '$lib/vocab/community/community';
@@ -14,7 +12,6 @@
 
 	const {
 		ui: {selectedSpace, spacesByCommunityId},
-		dispatch,
 	} = getApp();
 
 	export let persona: Readable<Persona>;
@@ -24,28 +21,6 @@
 	space; // TODO we're ignoring the space, but should probably mount its `content` as markup
 
 	$: communitySpaces = $spacesByCommunityId.get($community.community_id)!;
-
-	// TODO better way to do this? updates when switching communities,
-	// and updates only when actually changing the input --
-	// can this be simplfied drastically by not using `bind:value` on the hue?
-	let hue = $community.settings.hue;
-	let lastHue = hue;
-	let lastCommunity = community;
-	$: if (lastCommunity !== community) {
-		lastCommunity = community;
-		hue = $community.settings.hue;
-	}
-	$: if (hue !== lastHue) {
-		lastHue = hue;
-		updateCommunityHue(hue);
-	}
-	const UPDATE_INTERVAL = 500; // TODO extract this to config
-	const updateCommunityHue = throttle(UPDATE_INTERVAL, async (hue: number): Promise<void> => {
-		await dispatch('update_community_settings', {
-			community_id: $community.community_id,
-			settings: {hue},
-		});
-	});
 </script>
 
 <div class="markup">
@@ -80,21 +55,12 @@
 	</section>
 	<section>
 		<h2>settings</h2>
-		<!-- TODO maybe add a title or tooltip explaining `community.settings.hue` -->
-		<HueInput bind:hue />
-		<div class="community-icon">
-			<EntityIcon name={$community.name} type="Community" --hue={$community.settings.hue} />
-		</div>
+		<CommunityHue {community} />
 	</section>
 </div>
 
 <style>
 	section {
 		margin: var(--spacing_xl4) 0;
-	}
-	.community-icon {
-		/* TODO instead of this, maybe have a "centered-box" or "box" or "flex-centered" or copy Tailwind */
-		display: flex;
-		justify-content: center;
 	}
 </style>
